requirements.txt
aiofiles==0.5.0
aiohttp==3.6.2
aiopg==1.0.0
appdirs==1.4.4
async-timeout==3.0.1
attrs==19.3.0
black==19.10b0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.6.30
httptools==0.1.1
httpx==0.11.1
hyperframe==5.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.3.0
multidict==4.7.6
packaging==20.4
pathspec==0.8.0
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.1
pyparsing==2.4.7
pytest==5.4.3
PyYAML==5.3.1
regex==2020.6.8
rfc3986==1.4.0
sanic==20.6.3
six==1.15.0
sniffio==1.1.0
SQLAlchemy==1.3.18
toml==0.10.1
tornado==6.0.4
typed-ast==1.4.1
ujson==3.0.0
urllib3==1.25.9
uvloop==0.14.0
wcwidth==0.2.4
websockets==8.1
Werkzeug==1.0.1
yarl==1.4.2
